+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # "Barapost" changelog
 
+## 2019-11-19 edition (bug-fixing and performance update):
+
+### Noteworthy changes:
+
+1. barapost: "inc_lock 1-thr" bug fixed;
+2. barapost: rebuild-db "no -l no acc" bug fixed;
+3. prober: waiting for NCBI response no more makes a mess in terminal;
+
+### Version changes:
+
+1. prober: `1.12.k --> 1.12.l`;
+2. barapost: `3.6.a --> 3.6.b`;
+3. sorter: `3.4.a --> 3.4.b`;
+4. summarizer: `1.1.a --> 1.1.b`;
+
 ## 2019-11-17 edition:
 
-### Main changes:
+### Noteworthy changes:
 
 - sorter: singleFAST5 processing enabled;
 - barapost, sorter, summarizer: concise status bar embedded;